+++ src/main/java/mod/gottsch/forge/dungeonblocks/datagen/ItemModelsProvider.java
@@ -172,6 +172,7 @@ public class ItemModelsProvider extends ItemModelProvider {
 
 		blockItemParent(ModBlocks.MAP.get(ModBlocks.SQUARE_STONE_BRICK));
 		blockItemParent(ModBlocks.MAP.get(ModBlocks.MOSSY_SQUARE_STONE_BRICK));
+		blockItemParent(ModBlocks.MAP.get(ModBlocks.MOSSY_SQUARE_STONE_BRICK_STAIRS));
 		blockItemParent(ModBlocks.MAP.get(ModBlocks.SQUARE_MUD_BRICK));
 		// the facade / quarter facade item models are covered by the MAP sweep above
 		blockItemParent(ModBlocks.MAP.get(ModBlocks.SQUARE_STONE_BRICK_STAIRS));
@@ -187,6 +188,8 @@ public class ItemModelsProvider extends ItemModelProvider {
 		blockItemParent(ModBlocks.MAP.get(ModBlocks.MOSSY_RIGHT_LARGE_MUD_BRICK));
 		blockItemParent(ModBlocks.MAP.get(ModBlocks.POLISHED_ANDESITE_BRICKS));
 		blockItemParent(ModBlocks.MAP.get(ModBlocks.POLISHED_ANDESITE_BRICK_STAIRS));
+		blockItemParent(ModBlocks.MAP.get(ModBlocks.MOSSY_POLISHED_ANDESITE_BRICKS));
+		blockItemParent(ModBlocks.MAP.get(ModBlocks.MOSSY_POLISHED_ANDESITE_BRICK_STAIRS));
 		blockItemParent(ModBlocks.MAP.get(ModBlocks.MOSSY_BRICKS));
 		blockItemParent(ModBlocks.MAP.get(ModBlocks.MOSSY_BRICK_STAIRS));
 		blockItemParent(ModBlocks.MAP.get(ModBlocks.LARGE_BRICKS));